refactor(types): extract common type primitives
