+++ src/notebooklm/_types/__init__.py
@@ -1,0 +1,1 @@
+"""Private type implementation package."""
